requirements.txt
flask
flask_cors
w1thermsensor
pyalsaaudio
packaging
wireless